# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WhereWeAre API",
    "version": "1.0.0",
    "description": "Comprehensive API for accessing cultural landmarks, user management, memories, and favorites in Chemnitz"
  },
  "servers": [
    {
      "url": "https://whereweare.vercel.app",
      "description": "Application server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token"
      }
    },
    "schemas": {
...
      "Memory": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Unique identifier"
          },
          "userId": {
            "type": "string",
            "description": "User who created the memory"
          },
          "siteId": {
            "type": "string",
            "description": "Associated cultural site ID"
          },
          "siteName": {
            "type": "string",
            "description": "Name of the cultural site"
          },
          "siteCategory": {
            "type": "string",
            "description": "Category of the cultural site"
          },
          "coordinates": {
            "type": "object",
            "properties": {
              "lat": {
                "type": "number"
              },
              "lng": {
                "type": "number"
              }
            }
          },
          "title": {
            "type": "string",
            "description": "Optional title for the memory"
          },
          "note": {
            "type": "string",
            "description": "Memory description/note"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "filename": {
                  "type": "string"
                },
                "originalName": {
                  "type": "string"
                },
                "mimeType": {
                  "type": "string"
                },
                "size": {
                  "type": "number"
                }
              }
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isPublic": {
            "type": "boolean",
            "description": "Whether the memory is publicly visible"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "code": {
            "type": "string",
            "description": "Error code"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/memories": {
      "get": {
        "summary": "Get memories with various filters",
        "description": "Retrieve memories based on different filters. At least one filter parameter is required (userId, siteId, or public=true).",
        "tags": [
          "Memories"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "schema": {
              "type": "string"
            },
            "description": "Get memories for specific user"
          },
          {
            "in": "query",
            "name": "siteId",
            "schema": {
              "type": "string"
            },
            "description": "Get memories for specific site"
          },
          {
            "in": "query",
            "name": "public",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "description": "Get only public memories (use \"true\" to get public memories)"
          }
        ],
        "responses": {
          "200": {
            "description": "List of memories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Memory"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new memory",
        "description": "Create a new memory with images and metadata. All form fields are submitted as strings and parsed internally.",
        "tags": [
          "Memories"
        ],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "siteId",
                  "siteName",
                  "siteCategory",
                  "lat",
                  "lng",
                  "note",
                  "images"
                ],
                "properties": {
                  "siteId": {
                    "type": "string",
                    "description": "ID of the cultural site"
                  },
                  "siteName": {
                    "type": "string",
                    "description": "Name of the cultural site"
                  },
                  "siteCategory": {
                    "type": "string",
                    "description": "Category of the cultural site"
                  },
                  "lat": {
                    "type": "string",
                    "format": "float",
                    "description": "Latitude coordinate (parsed to number)"
                  },
                  "lng": {
                    "type": "string",
                    "format": "float",
                    "description": "Longitude coordinate (parsed to number)"
                  },
                  "title": {
                    "type": "string",
                    "description": "Optional title for the memory"
                  },
                  "note": {
                    "type": "string",
                    "description": "Memory description/note"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Image files (max 2MB each, 6MB total)"
                  },
                  "tags": {
                    "type": "string",
                    "description": "Comma-separated tags"
                  },
                  "isPublic": {
                    "type": "string",
                    "enum": [
                      "true",
                      "false"
                    ]
                  },
                  "description": "Whether memory is publicly visible"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Memory created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "memoryId": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Request entity too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}